backup: 2023-03-07 15:14
Affected files:
.obsidian/workspace.json
Narrative/1 - The Worst days of their lives/2. The Librarian's Accord.md
Narrative/1 - The Worst days of their lives/3..md
Narrative/1 - The Worst days of their lives/4. The forging of the minds.md
Narrative/1 - The Worst days of their lives/5. Upon which the party is founded.md
Narrative/1 - The Worst days of their lives/6. That time we did that thing.md