backup: 2023-03-15 22:35
Affected files:
.obsidian/workspace.json
Narrative/1 - The Worst days of their lives/3. Rise from the Ashes.md
Narrative/1 - The Worst days of their lives/4. That time we did that thing.md
Narrative/1 - The Worst days of their lives/5. The forging of the minds.md
Narrative/1 - The Worst days of their lives/6. The Auction.md
Narrative/1 - The Worst days of their lives/7. Upon which the party is founded.md
Narrative/Overviews/Currencies of The Empire.md